Légères corrections et améliorations
+ casts explicites plutôt qu'implicite
+ "||" au lieu de "or" dans debug.h
+ NO_DEBUG au lieu de la macro NDEBUG dans debug.h
+ get_road_width()  de parameters.h mis en commentaire, puisque non
utilisé jusqu'a présent
+ correction erreur de type dans FieldEvaluation::lengthBy()

+++ src/Engine/Coordinates.h
@@ -80,7 +80,7 @@ public:
 	* Calcul de la distance euclidienne entre deux points
 	* @return distance euclidienne : flottant non signé
 	*/
-    float euclideanDistance(const Coordinates& other) const;
+    double euclideanDistance(const Coordinates& other) const;
     
     
     /**
